--- tests/example_app/piccolo_migrations/2021-08-06T17-46-53-490098.py
@@ -1,178 +1,0 @@
-from piccolo.apps.migrations.auto import MigrationManager
-from piccolo.columns.column_types import UUID, Serial, Varchar
-from piccolo.columns.defaults.uuid import UUID4
-from piccolo.columns.indexes import IndexMethod
-
-ID = "2021-08-06T17:46:53:490098"
-VERSION = "0.29.0"
-
-
-async def forwards():
-    manager = MigrationManager(migration_id=ID, app_name="example_app")
-
-    manager.add_table("Album", tablename="album")
-
-    manager.add_column(
-        table_class_name="Album",
-        tablename="album",
-        column_name="id",
-        column_class_name="UUID",
-        column_class=UUID,
-        params={
-            "default": UUID4(),
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Album",
-        tablename="album",
-        column_name="name",
-        column_class_name="Varchar",
-        column_class=Varchar,
-        params={
-            "length": 255,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Manager",
-        tablename="manager",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Band",
-        tablename="band",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Venue",
-        tablename="venue",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Concert",
-        tablename="concert",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Ticket",
-        tablename="ticket",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Poster",
-        tablename="poster",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Shirt",
-        tablename="shirt",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="RecordingStudio",
-        tablename="recording_studio",
-        column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-        },
-    )
-
-    return manager
