+++ tests/example_app/piccolo_migrations/2021-08-08T13-17-33-948878.py
@@ -1,0 +1,434 @@
+from piccolo.apps.migrations.auto import MigrationManager
+from decimal import Decimal
+from enum import Enum
+from piccolo.columns.base import OnDelete
+from piccolo.columns.base import OnUpdate
+from piccolo.columns.column_types import ForeignKey
+from piccolo.columns.column_types import Integer
+from piccolo.columns.column_types import JSON
+from piccolo.columns.column_types import JSONB
+from piccolo.columns.column_types import Numeric
+from piccolo.columns.column_types import Serial
+from piccolo.columns.column_types import Text
+from piccolo.columns.column_types import UUID
+from piccolo.columns.column_types import Varchar
+from piccolo.columns.defaults.uuid import UUID4
+from piccolo.columns.indexes import IndexMethod
+from piccolo.table import Table
+
+
+class Band(Table, tablename="band"):
+    pass
+
+
+class Manager(Table, tablename="manager"):
+    pass
+
+
+class Venue(Table, tablename="venue"):
+    pass
+
+
+ID = "2021-08-08T13:17:33:948878"
+VERSION = "0.29.0"
+
+
+async def forwards():
+    manager = MigrationManager(migration_id=ID, app_name="example_app")
+
+    manager.add_table("Manager", tablename="manager")
+
+    manager.add_table("Venue", tablename="venue")
+
+    manager.add_table("Band", tablename="band")
+
+    manager.add_table("Concert", tablename="concert")
+
+    manager.add_table("Shirt", tablename="shirt")
+
+    manager.add_table("Ticket", tablename="ticket")
+
+    manager.add_table("Poster", tablename="poster")
+
+    manager.add_table("RecordingStudio", tablename="recording_studio")
+
+    manager.add_column(
+        table_class_name="Concert",
+        tablename="concert",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Concert",
+        tablename="concert",
+        column_name="band_1",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Band,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Concert",
+        tablename="concert",
+        column_name="band_2",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Band,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Concert",
+        tablename="concert",
+        column_name="venue",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Venue,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Manager",
+        tablename="manager",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Manager",
+        tablename="manager",
+        column_name="name",
+        column_class_name="Varchar",
+        column_class=Varchar,
+        params={
+            "length": 50,
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Venue",
+        tablename="venue",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Venue",
+        tablename="venue",
+        column_name="name",
+        column_class_name="Varchar",
+        column_class=Varchar,
+        params={
+            "length": 100,
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Venue",
+        tablename="venue",
+        column_name="capacity",
+        column_class_name="Integer",
+        column_class=Integer,
+        params={
+            "default": 0,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Band",
+        tablename="band",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Band",
+        tablename="band",
+        column_name="name",
+        column_class_name="Varchar",
+        column_class=Varchar,
+        params={
+            "length": 50,
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Band",
+        tablename="band",
+        column_name="manager",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Manager,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Band",
+        tablename="band",
+        column_name="popularity",
+        column_class_name="Integer",
+        column_class=Integer,
+        params={
+            "default": 0,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Shirt",
+        tablename="shirt",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Shirt",
+        tablename="shirt",
+        column_name="size",
+        column_class_name="Varchar",
+        column_class=Varchar,
+        params={
+            "length": 1,
+            "default": "l",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": Enum("Size", {"small": "s", "medium": "m", "large": "l"}),
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Ticket",
+        tablename="ticket",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Ticket",
+        tablename="ticket",
+        column_name="price",
+        column_class_name="Numeric",
+        column_class=Numeric,
+        params={
+            "default": Decimal("0"),
+            "digits": (5, 2),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Poster",
+        tablename="poster",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Poster",
+        tablename="poster",
+        column_name="content",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="RecordingStudio",
+        tablename="recording_studio",
+        column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="RecordingStudio",
+        tablename="recording_studio",
+        column_name="facilities",
+        column_class_name="JSON",
+        column_class=JSON,
+        params={
+            "default": "{}",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="RecordingStudio",
+        tablename="recording_studio",
+        column_name="facilities_b",
+        column_class_name="JSONB",
+        column_class=JSONB,
+        params={
+            "default": "{}",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+        },
+    )
+
+    return manager
